Collada loader moved to renderer.dll as well as Scene definitions
