+++ GLRenderer/GLRenderer/Mesh/StaticMeshResource.cpp
@@ -2,14 +2,14 @@
 
 #include <GLRenderer/Mesh/StaticMeshResource.h>
 
-#include <GLRenderer/MeshLoading/Scene.h>
+#include <Renderer/Mesh/Scene.h>
 
 namespace GLEngine {
 namespace GLRenderer {
 namespace Mesh {
 
 //=================================================================================
-C_StaticMeshResource::C_StaticMeshResource(const Mesh & mesh)
+C_StaticMeshResource::C_StaticMeshResource(const Renderer::MeshData::Mesh & mesh)
 {
 	static_assert(sizeof(glm::vec3) == sizeof(GLfloat) * 3, "Platform doesn't support this directly.");
