+++ GLRenderer/GLRenderer/MeshLoading/SceneBuilder.h
@@ -27,6 +27,13 @@ class xml_node;
 
 
 namespace GLEngine {
+namespace Renderer::MeshData
+{
+struct Scene;
+struct Mesh;
+struct Texture;
+}
+
 namespace GLRenderer {
 
 namespace Textures {
@@ -34,10 +41,6 @@ class C_Texture;
 }
 
 namespace Mesh {
-
-struct Scene;
-struct Mesh;
-struct Texture;
 class I_RenderNode;
 class C_Scene;
 class C_Terrain;
@@ -50,8 +53,8 @@ public:
 	std::shared_ptr<C_Scene> LoadScene(const std::string& sceneDefinitionFile);
 	std::shared_ptr<C_Terrain> LoadTerrain(const pugi::xml_node& node);
 	std::shared_ptr<C_Scene> LoadModel(const pugi::xml_node& node);
-	std::shared_ptr<I_RenderNode> LoadMesh(const Mesh& mesh);
-	std::shared_ptr<Textures::C_Texture> LoadTexture(const Texture& texture) const;
+	std::shared_ptr<I_RenderNode> LoadMesh(const Renderer::MeshData::Mesh& mesh);
+	std::shared_ptr<Textures::C_Texture> LoadTexture(const Renderer::MeshData::Texture& texture) const;
 private:
 	glm::vec3	ReadPositionNode(const pugi::xml_node& node) const noexcept;
 	std::string GetFolderpath(const std::string& filePath) const;
@@ -59,7 +62,7 @@ private:
 
 	std::vector<std::shared_ptr<Textures::C_Texture>> m_textures;
 	std::shared_ptr<Textures::C_Texture> m_nullTexture;
-	std::shared_ptr<Scene> m_scene;
+	std::shared_ptr<Renderer::MeshData::Scene> m_scene;
 	std::string m_sceneFolder;
 };
