+++ GLRenderer/GLRenderer/Textures/TextureLoader.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <GLRenderer/MeshLoading/Scene.h>
+#include <Renderer/Mesh/Scene.h>
 
 namespace GLEngine {
 namespace GLRenderer {
@@ -8,7 +8,7 @@ namespace Textures {
 class TextureLoader
 {
 public:
-	bool loadTexture(const char* path, Mesh::Texture& t);
+	bool loadTexture(const char* path, Renderer::MeshData::Texture& t);
 
 private:
 	static bool _isILinitialized;
